requirements.txt
fastapi[all]
pydantic
databases
sqlalchemy
uvicorn
alembic
psycopg2-binary
passlib[bcrypt]
python-jose